改进model获取器操作
