+++ tests/orm/ModelEventTest.php
@@ -5,7 +5,9 @@ namespace tests\orm;
 
 use PHPUnit\Framework\TestCase;
 use think\facade\Db;
-use think\Model;
+use tests\stubs\EventModel;
+use tests\stubs\EventObserver;
+use tests\stubs\ObservedModel;
 
 class ModelEventTest extends TestCase
 {
@@ -41,27 +43,8 @@ SQL
         $beforeInsertCalled = false;
         $afterInsertCalled = false;
 
-        $model = new class extends Model {
-            protected $table = 'test_event_model';
-            protected $autoWriteTimestamp = true;
-
-            public function onBeforeInsert($model)
-            {
-                global $beforeInsertCalled;
-                $beforeInsertCalled = true;
-                // 在插入前修改数据
-                $model->name = 'modified_' . $model->name;
-            }
-
-            public function onAfterInsert($model)
-            {
-                global $afterInsertCalled;
-                $afterInsertCalled = true;
-            }
-        };
-
         $data = ['name' => 'test3', 'status' => 1];
-        $result = $model::create($data);
+        $result = EventModel::create($data);
 
         $this->assertTrue($beforeInsertCalled, 'before_insert event not triggered');
         $this->assertTrue($afterInsertCalled, 'after_insert event not triggered');
@@ -73,27 +56,8 @@ SQL
         $beforeUpdateCalled = false;
         $afterUpdateCalled = false;
 
-        $model = new class extends Model {
-            protected $table = 'test_event_model';
-            protected $autoWriteTimestamp = true;
-
-            public function onBeforeUpdate($model)
-            {
-                global $beforeUpdateCalled;
-                $beforeUpdateCalled = true;
-                // 在更新前修改数据
-                $model->name = 'updated_' . $model->name;
-            }
-
-            public function onAfterUpdate($model)
-            {
-                global $afterUpdateCalled;
-                $afterUpdateCalled = true;
-            }
-        };
-
         // 先创建一条记录
-        $record = $model::create(['name' => 'test4', 'status' => 1]);
+        $record = EventModel::create(['name' => 'test4', 'status' => 1]);
 
         // 更新记录
         $record->name = 'new_name';
@@ -109,30 +73,9 @@ SQL
         $beforeDeleteCalled = false;
         $afterDeleteCalled = false;
 
-        $model = new class extends Model {
-            protected $table = 'test_event_model';
-            protected $autoWriteTimestamp = true;
-
-            public function onBeforeDelete($model)
-            {
-                global $beforeDeleteCalled;
-                $beforeDeleteCalled = true;
-                // 可以在删除前执行一些验证
-                if ($model->status === 0) {
-                    return false; // 阻止删除
-                }
-            }
-
-            public function onAfterDelete($model)
-            {
-                global $afterDeleteCalled;
-                $afterDeleteCalled = true;
-            }
-        };
-
         // 创建两条记录
-        $record1 = $model::create(['name' => 'test5', 'status' => 1]);
-        $record2 = $model::create(['name' => 'test6', 'status' => 0]);
+        $record1 = EventModel::create(['name' => 'test5', 'status' => 1]);
+        $record2 = EventModel::create(['name' => 'test6', 'status' => 0]);
 
         // 尝试删除状态为1的记录
         $record1->delete();
@@ -149,7 +92,7 @@ SQL
         $this->assertFalse($afterDeleteCalled, 'after_delete event should not be triggered');
 
         // 验证记录2仍然存在
-        $this->assertNotNull($model::find($record2->id));
+        $this->assertNotNull(EventModel::find($record2->id));
     }
 
     public function testWriteEvents()
@@ -157,27 +100,8 @@ SQL
         $beforeWriteCalled = false;
         $afterWriteCalled = false;
 
-        $model = new class extends Model {
-            protected $table = 'test_event_model';
-            protected $autoWriteTimestamp = true;
-
-            public function onBeforeWrite($model)
-            {
-                global $beforeWriteCalled;
-                $beforeWriteCalled = true;
-                // 在写入前修改数据
-                $model->name = 'write_' . $model->name;
-            }
-
-            public function onAfterWrite($model)
-            {
-                global $afterWriteCalled;
-                $afterWriteCalled = true;
-            }
-        };
-
         // 测试插入时的写入事件
-        $record = $model::create(['name' => 'test7', 'status' => 1]);
+        $record = EventModel::create(['name' => 'test7', 'status' => 1]);
         $this->assertTrue($beforeWriteCalled, 'before_write event not triggered on insert');
         $this->assertTrue($afterWriteCalled, 'after_write event not triggered on insert');
         $this->assertEquals('write_test7', $record->name);
@@ -196,74 +120,8 @@ SQL
 
     public function testModelObserver()
     {
-        $observer = new class {
-            public $beforeInsertCalled = false;
-            public $afterInsertCalled = false;
-            public $beforeUpdateCalled = false;
-            public $afterUpdateCalled = false;
-            public $beforeDeleteCalled = false;
-            public $afterDeleteCalled = false;
-            public $beforeWriteCalled = false;
-            public $afterWriteCalled = false;
-
-            public function onBeforeInsert($model)
-            {
-                $this->beforeInsertCalled = true;
-                $model->name = 'observer_' . $model->name;
-            }
-
-            public function onAfterInsert($model)
-            {
-                $this->afterInsertCalled = true;
-            }
-
-            public function onBeforeUpdate($model)
-            {
-                $this->beforeUpdateCalled = true;
-                $model->name = 'observer_updated_' . $model->name;
-            }
-
-            public function onAfterUpdate($model)
-            {
-                $this->afterUpdateCalled = true;
-            }
-
-            public function onBeforeDelete($model)
-            {
-                $this->beforeDeleteCalled = true;
-                if ($model->status === 0) {
-                    return false;
-                }
-            }
-
-            public function onAfterDelete($model)
-            {
-                $this->afterDeleteCalled = true;
-            }
-
-            public function onBeforeWrite($model)
-            {
-                $this->beforeWriteCalled = true;
-            }
-
-            public function onAfterWrite($model)
-            {
-                $this->afterWriteCalled = true;
-            }
-        };
-
-        $model = new class extends Model {
-            protected $table = 'test_event_model';
-            protected $autoWriteTimestamp = true;
-            protected $eventObserver;
-
-            public function __construct(array $data = [])
-            {
-                global $observer;
-                $this->eventObserver = $observer;
-                parent::__construct($data);
-            }
-        };
+        $observer = new EventObserver();
+        $model = new ObservedModel();
 
         // 测试插入事件
         $data = ['name' => 'test9', 'status' => 1];
